requirements.txt
pandasai
streamlit
python-dotenv
pandas